pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Lupin776 <Lupin776@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
customtkinter = "^5.2.2"
opencv-python = "^4.10.0.84"
pillow = "^11.0.0"
openai = "^1.55.2"
numpy = "^2.1.3"
scikit-learn = "^1.5.2"
joblib = "^1.4.2"
tiktoken = "^0.8.0"
imbalanced-learn = "^0.12.4"
matplotlib = "^3.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
